+++ src/__generated__/DetailTestsQuery.graphql.ts
@@ -87,7 +87,6 @@ fragment LocationMap_location on Location {
 
 fragment ShowHeader_show on Show {
   id
-  _id
   __id
   name
   press_release
@@ -265,42 +264,42 @@ v1 = {
 v2 = {
   "kind": "ScalarField",
   "alias": null,
-  "name": "id",
+  "name": "is_followed",
   "args": null,
   "storageKey": null
 },
 v3 = {
   "kind": "ScalarField",
   "alias": null,
-  "name": "city",
+  "name": "id",
   "args": null,
   "storageKey": null
 },
 v4 = {
   "kind": "ScalarField",
   "alias": null,
-  "name": "url",
+  "name": "city",
   "args": null,
   "storageKey": null
 },
 v5 = {
   "kind": "ScalarField",
   "alias": null,
-  "name": "aspect_ratio",
+  "name": "url",
   "args": null,
   "storageKey": null
 },
 v6 = {
   "kind": "ScalarField",
   "alias": null,
-  "name": "name",
+  "name": "aspect_ratio",
   "args": null,
   "storageKey": null
 },
 v7 = {
   "kind": "ScalarField",
   "alias": null,
-  "name": "is_followed",
+  "name": "name",
   "args": null,
   "storageKey": null
 },
@@ -319,7 +318,7 @@ v9 = {
   "storageKey": null
 },
 v10 = [
-  v6
+  v7
 ],
 v11 = {
   "kind": "InlineFragment",
@@ -350,14 +349,14 @@ v14 = [
   }
 ],
 v15 = [
-  v6,
+  v7,
   v1
 ];
 return {
   "kind": "Request",
   "operationKind": "query",
   "name": "DetailTestsQuery",
-  "id": "1bf8c40968856ceaa9f8b13344381fae",
+  "id": "eab855d93b250b4d45dcc71952390977",
   "text": null,
   "metadata": {},
   "fragment": {
@@ -400,14 +399,8 @@ return {
         "concreteType": "Show",
         "plural": false,
         "selections": [
-          {
-            "kind": "ScalarField",
-            "alias": null,
-            "name": "press_release",
-            "args": null,
-            "storageKey": null
-          },
           v2,
+          v3,
           {
             "kind": "ScalarField",
             "alias": null,
@@ -415,7 +408,7 @@ return {
             "args": null,
             "storageKey": null
           },
-          v3,
+          v4,
           {
             "kind": "LinkedField",
             "alias": null,
@@ -425,7 +418,7 @@ return {
             "concreteType": "Location",
             "plural": false,
             "selections": [
-              v2,
+              v3,
               {
                 "kind": "ScalarField",
                 "alias": null,
@@ -466,7 +459,7 @@ return {
                 ]
               },
               v1,
-              v3,
+              v4,
               {
                 "kind": "ScalarField",
                 "alias": null,
@@ -542,20 +535,19 @@ return {
             "concreteType": "Image",
             "plural": true,
             "selections": [
-              v2,
-              v4,
-              v5
+              v3,
+              v5,
+              v6
             ]
           },
+          v1,
           {
             "kind": "ScalarField",
             "alias": null,
-            "name": "_id",
+            "name": "press_release",
             "args": null,
             "storageKey": null
           },
-          v1,
-          v6,
           v7,
           v8,
           {
@@ -581,8 +573,8 @@ return {
                 "kind": "InlineFragment",
                 "type": "Partner",
                 "selections": [
-                  v6,
-                  v2,
+                  v7,
+                  v3,
                   {
                     "kind": "ScalarField",
                     "alias": null,
@@ -603,11 +595,11 @@ return {
             "concreteType": "Artist",
             "plural": true,
             "selections": [
-              v6,
+              v7,
               v12,
               v1,
+              v3,
               v2,
-              v7,
               {
                 "kind": "ScalarField",
                 "alias": null,
@@ -638,7 +630,7 @@ return {
                 "concreteType": "Image",
                 "plural": false,
                 "selections": [
-                  v4
+                  v5
                 ]
               }
             ]
@@ -676,7 +668,7 @@ return {
                 "concreteType": "Image",
                 "plural": false,
                 "selections": [
-                  v5,
+                  v6,
                   {
                     "kind": "ScalarField",
                     "alias": null,
@@ -721,7 +713,7 @@ return {
                 "args": null,
                 "storageKey": null
               },
-              v2,
+              v3,
               {
                 "kind": "ScalarField",
                 "alias": null,
@@ -912,8 +904,8 @@ return {
                     "plural": false,
                     "selections": [
                       v1,
-                      v2,
-                      v6,
+                      v3,
+                      v7,
                       v8,
                       {
                         "kind": "LinkedField",
@@ -924,8 +916,8 @@ return {
                         "concreteType": "Image",
                         "plural": true,
                         "selections": [
-                          v4,
-                          v5
+                          v5,
+                          v6
                         ]
                       },
                       {