+++ src/lib/Scenes/Fair/Components/FairHeader/index.tsx
@@ -38,11 +38,6 @@ const Overlay = styled.View`
   position: absolute;
 `
 
-const ButtonWrapper = styled(Box)`
-  width: 100%;
-  height: 85;
-`
-
 const Logo = styled(Image)`
   width: 150;
   height: 150;
@@ -50,6 +45,7 @@ const Logo = styled(Image)`
   margin-right: auto;
   background-color: transparent;
   margin-bottom: ${space(1)};
+  /* stylelint-disable */
   tint-color: white;
 `
 
@@ -191,8 +187,8 @@ export class FairHeader extends React.Component<Props, State> {
         <Spacer mt={2} />
         <Box mx={2}>{this.getContextualDetails()}</Box>
         {profile && (
-          <ButtonWrapper px={2}>
-            <Spacer m={1} mt={1} />
+          <Box px={2}>
+            <Spacer m={2} mt={1} />
             <InvertedButton
               text={profile.is_followed ? "Fair saved" : "Save fair"}
               onPress={() => this.handleSaveFair()}
@@ -202,7 +198,7 @@ export class FairHeader extends React.Component<Props, State> {
               buttonSize={"large"}
             />
             <Spacer m={1} />
-          </ButtonWrapper>
+          </Box>
         )}
       </>
     )