+++ CHANGELOG.md
@@ -21,6 +21,10 @@
 - Moves countdown timer higher on fair - kieran
 - Fixes space under map when no description - kieran
 - Fixes show title header when no carousel images - kieran
+- Refactors Buckets for All events city view - luc
+- Adds Fair Section in city view - luc
+- Adds AllEvents data and styling to city view - kieran
+- Adds Saved Shows section to city view - luc
 - Styles Fair save button - kieran
 - Fixes save show - kieran
 - Show header title now links to gallery - kieran